\name{asmbPLS-package}
\alias{asmbPLS-package}
\alias{asmbPLS}
\docType{package}
\title{
\packageTitle{asmbPLS}
}
\description{
\packageDescription{asmbPLS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{asmbPLS}
\packageIndices{asmbPLS}
}
\author{
\packageAuthor{asmbPLS}

Maintainer: \packageMaintainer{asmbPLS}
}
\references{
add later
}
\seealso{
\code{\link{asmbPLS.fit}}, \code{\link{asmbPLS.cv}}, \code{\link{asmbPLS.predict}}, \code{\link{mbPLS.fit}}, \code{\link{meanimp}}
}

