% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmbPLS.fit.R
\name{asmbPLS.fit}
\alias{asmbPLS.fit}
\title{asmbPLS for block-structured data}
\usage{
asmbPLS.fit(X.matrix, Y.matrix, PLS.comp, X.dim, quantile.comb)
}
\arguments{
\item{X.matrix}{Predictors matrix. Samples in rows, variables in columns}

\item{Y.matrix}{Outcome matrix. Samples in rows, this is a matrix with one 
column (continuous variable). The outcome could be imputed survival time. 
For survival time with right-censored survival time and event indicator, the 
right censored time could be imputed by \code{\link{meanimp}}.}

\item{PLS.comp}{Number of PLS components in asmbPLS.}

\item{X.dim}{A vector containing the number of predictors in each block 
(ordered).}

\item{quantile.comb}{A matrix containing quantile combinations used for 
different PLS components, whose row number equals to the number of PLS 
components used, column number equals to the number of blocks.}
}
\value{
\code{asmbPLS.fit} returns a list containing the following components:
\item{X_dim}{A vector containing the number of predictors in each block.}
\item{x_weight}{A list containing the weights of predictors for different 
blocks in different PLS components.}
\item{x_score}{A list containing the scores of samples in different blocks
in different PLS components.}
\item{x_loading}{A list containing the loadings of predictors for different
blocks in different PLS components.}
\item{x_super_weight}{A matrix containing the super weights of different
blocks for different PLS components.}
\item{x_super_score}{A matrix containing the super scores of samples for
different PLS components.}
\item{y_weight}{A matrix containing the weights of outcome for different 
PLS components.}
\item{y_score}{A matrix containing the scores of outcome for different 
PLS components.}
\item{col_mean}{A vector containing the mean of each predictor for scaling.}
\item{col_sd}{A vector containing the standard deviation of each predictor
for scaling. Predictor with sd = 0 will be changed to 1.}
\item{X_scaled}{Scaled predictors matrix.}
\item{Y_mean}{The mean of outcome matrix for scaling.}
\item{Y_sd}{The standard deviation of outcome matrix for scaling.}
\item{Y_scaled}{Scaled outcome matrix.}
}
\description{
Function to fit the adaptive sparse multi-block partial least square model 
(asmbPLS) with several explanatory blocks (X_1, ..., X_B) as our predictors
to explain the outcome Y.
}
\examples{
## Use the example dataset
data(asmbPLS.fit.example)
 
## asmbPLS fit
asmbPLS.results <- asmbPLS.fit(X.matrix = asmbPLS.fit.example$X.matrix, 
Y.matrix = asmbPLS.fit.example$Y.matrix, 
PLS.comp = asmbPLS.fit.example$PLS.comp, 
X.dim = asmbPLS.fit.example$X.dim, 
quantile.comb = asmbPLS.fit.example$quantile.comb)

}
