% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmbPLSDA.cv.example.R
\docType{data}
\name{asmbPLSDA.cv.example}
\alias{asmbPLSDA.cv.example}
\title{Example data for asmbPLSDA CV}
\format{
A list including 6 components:  
1) X.matrix, a matrix with 100 samples (rows) and 400 features; 
2) Y.matrix.binary, a matrix with 100 samples (rows) and 1 column; 
3) Y.matrix.morethan2levels, a matrix with 100 samples (rows) and 3 columns (3 levels); 
4) X.dim, dimension of the two blocks in X.matrix; 
5) quantile.comp, selected quantile combinations; 
6) PLS.comp, selected number of PLS components
}
\usage{
data(asmbPLSDA.cv.example)
}
\description{
Simulated data for asmbPLSDA CV.
}
\keyword{datasets}
