% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmbPLSDA.predict.example.R
\docType{data}
\name{asmbPLSDA.predict.example}
\alias{asmbPLSDA.predict.example}
\title{Example data for asmbPLS-DA predict}
\format{
A list including 7 components:  
1) X.matrix, a matrix with 100 samples (rows) and 400 features; 
2) Y.matrix.binary, a matrix with 100 samples (rows) and 1 column; 
3) Y.matrix.morethan2levels, a matrix with 100 samples (rows) and 3 columns (3 levels);
4) X.dim, dimension of the two blocks in X.matrix; 
5) quantile.comp, selected quantile combinations; 
6) PLS.comp, selected number of PLS components
7) X.matrix.new, a matrix to be predicted with 100 samples (rows) and 400 features;
}
\usage{
data(asmbPLSDA.predict.example)
}
\description{
Simulated data for asmbPLS-DA predict.
}
\keyword{datasets}
