% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.categorical.R
\name{to.categorical}
\alias{to.categorical}
\title{Converts a class vector to binary class matrix}
\usage{
to.categorical(data)
}
\arguments{
\item{categorical.vector}{A class vector}
}
\value{
\item{output.matrix}{The output binary class matrix, which can be used for asmbPLS-DA.}
}
\description{
Converts a class vector to binary class matrix with the number of columns equal to the number of levels
}
\examples{
## Generate a class vector
vector.test <- factor(c(1,1,1,2,2,2,3,3,3), levels = c(1,2,3))

## convert
output.matrix <- to.categotical(vector.test)

}
