\name{asmbPLS-package}
\alias{asmbPLS-package}
\alias{asmbPLS}
\docType{package}
\title{
\packageTitle{asmbPLS}
}
\description{
\packageDescription{asmbPLS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{asmbPLS}
\packageIndices{asmbPLS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{asmbPLS}

Maintainer: \packageMaintainer{asmbPLS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions88 ~~
}
