% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmbPLS.cv.R
\name{asmbPLS.cv}
\alias{asmbPLS.cv}
\title{Cross-validation for asmbPLS to find the best combinations of quantiles for prediction}
\usage{
asmbPLS.cv(
  X.matrix,
  Y.matrix,
  PLS.comp,
  X.dim,
  quantile.comb.table,
  k = 5,
  Y.indicator,
  only.observe = T
)
}
\arguments{
\item{X.matrix}{Predictors matrix. Samples in rows, variables in columns}

\item{Y.matrix}{Outcome matrix. Samples in rows, this is a matrix with one 
column (continuous variable). The outcome could be imputed survival time. 
For survival time with right-censored survival time and event indicator, the 
right censored time could be imputed by \code{\link{meanimp}}.}

\item{PLS.comp}{Number of PLS components in asmbPLS.}

\item{X.dim}{A vector containing the number of predictors in each block 
(ordered).}

\item{quantile.comb.table}{A matrix containing user-defined quantile 
combinations used for CV, whose column number equals to the 
number of blocks.}

\item{k}{The number of folds of CV procedure. The default is 5.}

\item{Y.indicator}{A vector containing the event indicator for each sample, 
whose length is equal to the number of samples. This vector allows the ratio 
of observed/unobserved to be the same in the training set and validation set. 
Observed = 1, and unobserved = 0. If other types of outcome data rather than 
survival outcome is used, you can use a vector with all components = 1 
instead.}

\item{only.observe}{Whether only observed samples in the validation set 
should be used for calculating the MSE for CV. The default is TRUE.}
}
\value{
\code{asmbPLS.cv} returns a list containing the following components:
\item{quantile.comb}{A matrix containing the selected quantile 
combination and the corresponding MSE of CV for each PLS component.}
\item{time}{Computation time.}
}
\description{
Function to find the best combinations of quantiles used for prediction via
cross-validation. Usually should be conducted before 
\code{\link[asmbPLS]{asmbPLS.fit}} to obtain the quantile combinations.
}
\examples{
## Use the example dataset
data(asmbPLS.cv.example)

## cv to find the best quantile combinations for model fitting
cv.results <- asmbPLS.cv(X.matrix = X.matrix, Y.matrix = Y.matrix, PLS.comp 
= PLS.comp, X.dim = X.dim, quantile.comb.table = quantile.comb.table, k = 5, 
Y.indicator = Y.indicator, only.observe = T)
quantile.comb <- cv.results$quantile.comb[,1:2]
 
## asmbPLS fit
asmbPLS.results <- asmbPLS.fit(X.matrix = X.matrix, Y.matrix = Y.matrix, 
PLS.comp = PLS.comp, X.dim = X.dim, quantile.comb = quantile.comb)

}
