% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbPLS.fit.R
\name{mbPLS.fit}
\alias{mbPLS.fit}
\title{mbPLS for block-structured data}
\usage{
mbPLS.fit(X.matrix, Y.matrix, PLS.comp, X.dim)
}
\arguments{
\item{X.matrix}{Predictors matrix. Samples in rows, variables in columns}

\item{Y.matrix}{Outcome matrix. Samples in rows, this is a matrix with one 
column (continuous variable). The outcome could be imputed survival time. 
For survival time with right-censored survival time and event indicator, the 
right censored time could be imputed by \code{\link{meanimp}}.}

\item{PLS.comp}{Number of PLS components in mbPLS.}

\item{X.dim}{A vector containing the number of predictors in each block 
(ordered).}
}
\value{
\code{mbPLS.fit} returns a list containing the following components:
\item{X_dim}{A vector containing the number of predictors in each block.}
\item{x_weight}{A list containing the weights of predictors for different 
blocks in different PLS components.}
\item{x_score}{A list containing the scores of samples in different blocks
in different PLS components.}
\item{x_loading}{A list containing the loadings of predictors for different
blocks in different PLS components.}
\item{x_super_weight}{A matrix containing the super weights of different
blocks for different PLS components.}
\item{x_super_score}{A matrix containing the super scores of samples for
different PLS components.}
\item{y_weight}{A matrix containing the weights of outcome for different 
PLS components.}
\item{y_score}{A matrix containing the scores of outcome for different 
PLS components.}
\item{col_mean}{A vector containing the mean of each predictor for scaling.}
\item{col_sd}{A vector containing the standard deviation of each predictor
for scaling. Predictor with sd = 0 will be changed to 1.}
\item{X_scaled}{Scaled predictors matrix.}
\item{Y_mean}{The mean of outcome matrix for scaling.}
\item{Y_sd}{The standard deviation of outcome matrix for scaling.}
\item{Y_scaled}{Scaled outcome matrix.}
}
\description{
Function to fit the multi-block partial least square model 
(mbPLS) with several explanatory blocks (X_1, ..., X_B) as our predictors
to explain the outcome Y.
}
\examples{
## Use the example dataset
data(asmbPLS.fit.example)
 
## mbPLS fit
mbPLS.results <- mbPLS.fit(X.matrix = X.matrix, Y.matrix = Y.matrix, 
PLS.comp = PLS.comp, X.dim = X.dim)

}
